TLight/requirements.txt
asgiref==3.5.0
Django==4.0.1
django-debug-toolbar==3.2.4
django-seed==0.3.1
djangorestframework==3.13.1
Faker==11.3.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toposort==1.7
